+++ rust/src/interpreter/mod.rs
@@ -1,18 +1,21 @@
 use std::{cmp::min, mem, slice};
 
 use evmc_vm::{
-    AccessStatus, Address, ExecutionContext, ExecutionMessage, MessageFlags, MessageKind, Revision,
+    AccessStatus, ExecutionContext, ExecutionMessage, MessageFlags, MessageKind, Revision,
     StatusCode, StepResult, StepStatusCode, StorageStatus, Uint256,
 };
 use sha3::{Digest, Keccak256};
 
 use crate::{
-    interpreter::memory::Memory,
+    interpreter::{checks::*, gas::*, memory::Memory, stack::*},
     types::{opcode, u256},
 };
 
+mod checks;
 mod code_state;
+mod gas;
 mod memory;
+mod stack;
 
 pub use code_state::CodeState;
 
@@ -1357,159 +1360,6 @@ fn static_delegate_call<const DELEGATE: bool>(
     stack.push(((result.status_code() == StatusCode::EVMC_SUCCESS) as u8).into());
     code_state.next();
     Ok(())
-}
-
-#[inline(always)]
-fn check_min_revision(
-    min_revision: Revision,
-    revision: Revision,
-) -> Result<(), (StepStatusCode, StatusCode)> {
-    if revision < min_revision {
-        return Err((
-            StepStatusCode::EVMC_STEP_FAILED,
-            StatusCode::EVMC_UNDEFINED_INSTRUCTION,
-        ));
-    }
-    Ok(())
-}
-
-#[inline(always)]
-fn consume_gas<const GAS: u64>(gas_left: &mut u64) -> Result<(), (StepStatusCode, StatusCode)> {
-    if *gas_left < GAS {
-        return Err((
-            StepStatusCode::EVMC_STEP_FAILED,
-            StatusCode::EVMC_OUT_OF_GAS,
-        ));
-    }
-    *gas_left -= GAS;
-    Ok(())
-}
-
-#[inline(always)]
-fn consume_dyn_gas(gas_left: &mut u64, gas: u64) -> Result<(), (StepStatusCode, StatusCode)> {
-    if *gas_left < gas {
-        return Err((
-            StepStatusCode::EVMC_STEP_FAILED,
-            StatusCode::EVMC_OUT_OF_GAS,
-        ));
-    }
-    *gas_left -= gas;
-    Ok(())
-}
-
-#[inline(always)]
-fn consume_positive_value_cost(
-    value: &u256,
-    gas_left: &mut u64,
-) -> Result<(), (StepStatusCode, StatusCode)> {
-    if *value != u256::ZERO {
-        consume_gas::<9000>(gas_left)?;
-    }
-    Ok(())
-}
-
-#[inline(always)]
-fn consume_value_to_empty_account_cost(
-    value: &u256,
-    addr: &Address,
-    context: &mut ExecutionContext,
-    gas_left: &mut u64,
-) -> Result<(), (StepStatusCode, StatusCode)> {
-    if *value != u256::ZERO && !context.account_exists(&addr) {
-        consume_gas::<25000>(gas_left)?;
-    }
-    Ok(())
-}
-
-#[inline(always)]
-fn consume_address_access_cost(
-    gas_left: &mut u64,
-    addr: &Address,
-    context: &mut ExecutionContext,
-    revision: Revision,
-) -> Result<(), (StepStatusCode, StatusCode)> {
-    let tx_context = context.get_tx_context();
-    if revision >= Revision::EVMC_BERLIN {
-        if *addr != tx_context.tx_origin
-            //&& addr != tx_context.tx_to // TODO
-            && !(revision >= Revision::EVMC_SHANGHAI && *addr == tx_context.block_coinbase)
-            && context.access_account(addr) == AccessStatus::EVMC_ACCESS_COLD
-        {
-            consume_gas::<2600>(gas_left)?;
-        } else {
-            consume_gas::<100>(gas_left)?;
-        }
-    }
-    Ok(())
-}
-
-/// consume 3 * minimum_word_size
-#[inline(always)]
-fn consume_copy_cost(gas_left: &mut u64, len: u64) -> Result<(), (StepStatusCode, StatusCode)> {
-    consume_dyn_gas(gas_left, 3 * word_size(len))?;
-    Ok(())
-}
-
-#[inline(always)]
-fn check_stack_overflow<const N: usize>(
-    stack: &[u256],
-) -> Result<(), (StepStatusCode, StatusCode)> {
-    if stack.len() + N > 1024 {
-        return Err((
-            StepStatusCode::EVMC_STEP_FAILED,
-            StatusCode::EVMC_STACK_OVERFLOW,
-        ));
-    }
-    Ok(())
-}
-
-#[inline(always)]
-fn check_not_read_only(
-    message: &ExecutionMessage,
-    revision: Revision,
-) -> Result<(), (StepStatusCode, StatusCode)> {
-    if revision >= Revision::EVMC_BYZANTIUM {
-        let read_only = message.flags() == MessageFlags::EVMC_STATIC as u32;
-        if read_only {
-            return Err((
-                StepStatusCode::EVMC_STEP_FAILED,
-                StatusCode::EVMC_STATIC_MODE_VIOLATION,
-            ));
-        }
-    }
-    Ok(())
-}
-
-#[inline(always)]
-fn pop_from_stack<const N: usize>(
-    stack: &mut Vec<u256>,
-) -> Result<[u256; N], (StepStatusCode, StatusCode)> {
-    if stack.len() < N {
-        return Err((
-            StepStatusCode::EVMC_STEP_FAILED,
-            StatusCode::EVMC_STACK_UNDERFLOW,
-        ));
-    }
-    let mut array = [u256::ZERO; N];
-    for element in &mut array {
-        *element = stack.pop().unwrap();
-    }
-
-    Ok(array)
-}
-
-#[inline(always)]
-fn nth_ref_from_stack<const N: usize>(
-    stack: &[u256],
-) -> Result<&u256, (StepStatusCode, StatusCode)> {
-    if stack.len() < N {
-        return Err((
-            StepStatusCode::EVMC_STEP_FAILED,
-            StatusCode::EVMC_STACK_UNDERFLOW,
-        ));
-    }
-
-    Ok(&stack[stack.len() - N])
 }
 
 fn get_slice_within_bounds<T>(data: &[T], offset: u256, len: u64) -> &[T] {
